+++ payloads/v2/logs.ts
@@ -1,7 +1,7 @@
 import type { APIPayload } from "../../common/v2";
 
 /**
- * @see https://docs.squarecloud.app/api-reference/endpoint/apps/logs
+ * @see https://docs.squarecloud.app/en/api-reference/endpoint/apps/logs
  */
 export interface APIApplicationLogs {
 	logs: string;